requirements.txt
fastai==2.7.12
torch==2.1.0
flask==3.0.0
pillow==10.1.0
numpy==1.24.3 